createVoviesWeek
