+++ src/index.html
@@ -12,7 +12,7 @@
   <body>
     <include src="./partials/hero.html"></include>
     <include src="./partials/header.html"></include>
-    <div class="gallary-films"></div>
+    <include src="./partials/main-movies.html"></include>
     <include src="./partials/footer.html"></include>
     <include src="./partials/modal-review.html"></include>
     <include src="./partials/modal-sign-in.html"></include>
