+++ src/sass/index.scss
@@ -13,42 +13,8 @@
 @import './components/pagination';
 
 @import './header';
+@import './main-movies';
 @import './hero.scss';
 @import './footer';
 
 
-.gallary-films {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;    
-}
-img { 
-    width: 480px;   
-    display: block;
-    width: 100%;
-    height: 200px;
-    border-top-left-radius: 8px;
-    border-top-right-radius: 8px;
-}
-// .photo-card {    
-//     width: 320px;
-//     margin-right: 10px;
-//     margin-bottom: 10px;
-//     background: white;
-//     border: 1px solid #cddad0;
-//     box-sizing: border-box;
-//     border-radius: 8px;
-//     background-color: rgb(207, 241, 230);
-// }
-// a { 
-//     display: flex;
-//     justify-content: space-around;
-//     text-decoration: none;
-// }
-// .info-item {
-//     text-align: center;
-//     font-size: 12px;
-//     line-height: 1.40;
-//     color:black;
-//     margin: 0px;
-// }
